* This is a work-around for a completely broken implementation. Specifically, thefunction call stack looks like:
    fcgi_request_service<Protocol>::handle_begin_request_header ->
    fcgi_request_service<Protocol>::parse_post_vars ->
    request_base::parse_post_vars ->
    form_parser::parse

Form parser explicitly states in documentation that valid form encodings are application/x-www-form-urlencoded, and multipart/form-data.  I am not familiar enough with HTML forms to know if this restriction is legitamite for html forms or not.  However, it is an invalid restriction for HTML POST in general, as long as the mime type is at least IANA registered (http://www.iana.org/assignments/media-types).

The code change is to allow POST of any (presumably IANA registered) mime type.  If the specific service implementation doesn't want to except it, then it can always return an error, but the library aught not prevent the server from accepting valid mime types, even if the server implementation hast to parse it manually.

(This problem was discovered when attempting to implement a RESTful service which accepts GET/POST for application/json)